refactor: валидация дат booking перенесена из server в gateway
